+++ packages/misc-util/src/net/http/_http-fields.test.ts
@@ -1,6 +1,6 @@
 import { inspect } from "node:util";
 import { expect, suite, test } from "vitest";
-import { HttpFields } from "./http-fields.js";
+import { HttpFields } from "./_http-fields.js";
 
 suite("HttpHeaders", () => {
 	test("should create empty fields", () => {